SafeTraversal.jl

Documentation for SafeTraversal.jl

SafeTraversal.:⋄ — Function

key ⋄ object

Returns object[key] if object is not nothing or missing andkeyis a valid key inobject, otherwise returnsnothing`.

Examples

julia> nothing ⋄ :a

julia> Dict(:a => 1) ⋄ :a
1

julia> Dict(:a => 1) ⋄ :b

SafeTraversal.@⋄ — Macro

@⋄ var1=expr evaluation
@⋄ var1 evaluation

Evaluates the first provided expression, and if it is not nothing or missing, evaluates the second expression using the first expression as a local variable. For instance:

@⋄ x=[1,2,3] x[1]

This will evaluate to 1, and is equivalent to

let x = [1,2,3]
    if isnothing(x) || ismissing(x)
        nothing
    else
        x[1]
    end
end

If the first expression is a single symbol, the value of that variable is checked for null:

julia> x = [1,2,3]
3-element Vector{Int64}:
 1
 2
 3

julia> @⋄ x x[1]
1